Move content from PersonSearchActivity to PersonSearchFragment
