+++ app/src/main/java/de/tum/in/tumcampusapp/component/tumui/person/PersonSearchResultsAdapter.kt
@@ -1,16 +1,16 @@
 package de.tum.`in`.tumcampusapp.component.tumui.person
 
-import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
 import de.tum.`in`.tumcampusapp.R
 import de.tum.`in`.tumcampusapp.component.tumui.person.model.Person
-import kotlinx.android.synthetic.main.person_search_result_item.view.*
+import kotlinx.android.synthetic.main.person_search_result_item.view.textView
 
 class PersonSearchResultsAdapter(
         private var items: List<Person>,
-        private val listener: PersonSearchResultsItemListener
+        private val onItemClick: (Person) -> Unit
 ) : RecyclerView.Adapter<PersonSearchResultsAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -20,7 +20,7 @@ class PersonSearchResultsAdapter(
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bind(items[position], listener)
+        holder.bind(items[position], onItemClick)
     }
 
     override fun getItemCount() = items.size
@@ -32,9 +32,12 @@ class PersonSearchResultsAdapter(
 
     class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
-        fun bind(person: Person, listener: PersonSearchResultsItemListener) = with(itemView) {
+        fun bind(
+            person: Person,
+            onItemClick: (Person) -> Unit
+        ) = with(itemView) {
             textView.text = person.getFullName()
-            setOnClickListener { listener.onItemSelected(person) }
+            setOnClickListener { onItemClick(person) }
         }
 
     }
